Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int a=scanf(int)
+    float a=scanf(float)
     printf("You entered: %s", a);
     return 0;
 }
